Move the pan dragger to the RootPanel.
In pan mode, the RootPanel still allows child panels to override the
drag behavior, but now the default panning is handled where all the
dragging behavior is handled.

As a side effect, this fixes the issue where panning was not working on
title panels.
